requirements.txt
biopython==1.66
cycler==0.9.0
Cython==0.23.4
decorator==4.0.6
ipython==4.0.3
ipython-genutils==0.1.0
matplotlib==1.5.1
nose==1.3.7
numpy==1.10.2
pandas==0.17.1
path.py==0.0.0
pexpect==3.3
pickleshare==0.5
primer3-py==0.4.3
py==1.4.31
pybedtools==0.7.4
pycairo==1.10.0
pycrispr==0.1.2
pyparsing==2.0.3
pysam==0.8.4
pytest==2.8.5
python-dateutil==2.4.2
pytz==2015.7
simplegeneric==0.8.1
six==1.10.0
traitlets==4.1.0
wheel==0.26.0
